refactor(host-application): centralize app_service test support and split integrations
Move shared app_service test scaffolding into test_support and keep tests/mod.rs as module wiring only.

Replace the monolithic integration test file with focused integration submodules and keep unit tests on explicit imports.
