+++ apps/desktop/src-tauri/crates/host-application/src/app_service/tests/integration/mod.rs
@@ -1,0 +1,6 @@
+mod build_flows;
+mod git_and_task;
+mod opencode_runtime;
+mod runtime_checks;
+mod shutdown_and_helpers;
+mod workflow_docs;
